# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfólio API Docs",
        "description": "API para consulta de dados do portfólio",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "Licença API",
            "url": "https://github.com/felipesilva15/portfolio-manager-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/project-type": {
            "get": {
                "tags": [
                    "ProjectType"
                ],
                "summary": "List all project types",
                "operationId": "519b01eb81768fe50ddafb980133a0d9",
                "responses": {
                    "200": {
                        "description": "Project Type list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectType"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ProjectType"
                ],
                "summary": "Registers a project type",
                "operationId": "0b1a543b1c348880c24115ba33613232",
                "requestBody": {
                    "description": "Data for creating a new project type",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered project type data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectType"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/project-type/{id}": {
            "get": {
                "tags": [
                    "ProjectType"
                ],
                "summary": "List a project type by ID",
                "operationId": "6a54dbdc9a82492917c5bb51250fab89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Type data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectType"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ProjectType"
                ],
                "summary": "Update a project type",
                "operationId": "4b0e62dfa27b4f7c23d6772adddd21e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for update project type",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated project type data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectType"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProjectType"
                ],
                "summary": "Delete a project type",
                "operationId": "74d63c88342e2af173daf9ac131b036f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiErrorDTO": {
                "properties": {
                    "path": {
                        "type": "string",
                        "example": "/api/route"
                    },
                    "code": {
                        "type": "integer",
                        "example": 500
                    },
                    "message": {
                        "type": "string",
                        "example": "Error ocurried"
                    }
                },
                "type": "object"
            },
...
            "ProjectTypeRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 40,
                        "minLength": 1,
                        "example": "Back-end"
                    }
                },
                "type": "object"
            },
...
            "ProjectType": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Back-end"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T11:23:31.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T11:23:31.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}